backend/requirements.txt
black==23.11.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Django==4.2.7
drf-yasg==1.21.7
googlemaps==4.10.0
Pillow==10.1.0
psycopg2==2.9.9
pyjwt==2.8.0
redis==5.0.1
sendgrid==6.11.0
playwright